{{cookiecutter.package_name}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.package_name}}"
version = "{{cookiecutter.package_version}}"
description = "{{cookiecutter.project_description}}"
authors = ["{{cookiecutter.author_name}} <{{cookiecutter.author_email}}>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^3.0.0"
coverage-badge = "^1.0.2"
nox = "^2021.10.1"
poetry-core = "^1.0.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
